final push

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -1,6 +1,7 @@
 
 package inventory.controller;
 
+import inventory.logger.ApplicationLogger;
 import inventory.model.Part;
 import inventory.model.Product;
 import inventory.service.InventoryService;
@@ -20,23 +21,19 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
+import java.util.logging.Level;
 
 
 public class MainScreenController implements Initializable,Controller {
-    
-     // Declare fields
-    private Stage stage;
-    private Parent scene;
-    private static Part modifyPart;
-    private static Product modifyProduct;
+
     private static int modifyPartIndex;
     private static int modifyProductIndex;
-    
+
     // Declare methods
     public static int getModifyPartIndex() {
         return modifyPartIndex;
     }
-    
+
     public static int getModifyProductIndex() {
         return modifyProductIndex;
     }
@@ -73,14 +70,13 @@ public class MainScreenController implements Initializable,Controller {
 
     @FXML
     private TableColumn<Product, Double> productsPriceCol;
-    
+
     @FXML
     private TextField partsSearchTxt;
-    
+
     @FXML
     private TextField productsSearchTxt;
 
-    public MainScreenController(){}
 
     public void setService(InventoryService service){
         this.service=service;
@@ -114,10 +110,10 @@ public class MainScreenController implements Initializable,Controller {
      * @throws IOException
      */
     private void displayScene(ActionEvent event, String source) throws IOException {
-        stage = (Stage)((Button)event.getSource()).getScene().getWindow();
+        // Declare fields
+        Stage stage = (Stage) ((Button) event.getSource()).getScene().getWindow();
         FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
-        //scene = FXMLLoader.load(getClass().getResource(source));
-        scene = loader.load();
+        Parent scene = loader.load();
         Controller ctrl=loader.getController();
         ctrl.setService(service);
         stage.setScene(new Scene(scene));
@@ -126,7 +122,7 @@ public class MainScreenController implements Initializable,Controller {
 
     /**
      * Ask user for confirmation before deleting selected part from list of parts.
-     * @param event 
+     * @param event
      */
     @FXML
     void handleDeletePart(ActionEvent event) {
@@ -139,34 +135,35 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            ApplicationLogger.log(Level.INFO, "Part deleted");
             service.deletePart(part);
         } else {
-            System.out.println("Canceled part deletion.");
+            ApplicationLogger.log(Level.INFO,"Canceled part deleteion.");
         }
     }
 
     /**
      * Ask user for confirmation before deleting selected product from list of products.
-     * @param event 
+     * @param event
      */
     @FXML
     void handleDeleteProduct(ActionEvent event) {
         Product product = productsTableView.getSelectionModel().getSelectedItem();
-        
+
         Alert alert = new Alert(AlertType.CONFIRMATION);
         alert.initModality(Modality.NONE);
         alert.setTitle("Confirmation");
         alert.setHeaderText("Confirm Product Deletion?");
         alert.setContentText("Are you sure you want to delete product " + product.getName() + " from products?");
         Optional<ButtonType> result = alert.showAndWait();
-        
-        if (result.get() == ButtonType.OK) {
+
+        if (result.isPresent() && result.get() == ButtonType.OK) {
             service.deleteProduct(product);
-            System.out.println("Product " + product.getName() + " was removed.");
+            ApplicationLogger.log(Level.INFO,"Product " + product.getName() + " was removed.");
         } else {
-            System.out.println("Product " + product.getName() + " was not removed.");
+            ApplicationLogger.log(Level.INFO,"Product " + product.getName() + " was not removed.");
+
         }
     }
 
@@ -198,9 +195,9 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyPart(ActionEvent event) throws IOException {
-        modifyPart = partsTableView.getSelectionModel().getSelectedItem();
+        Part modifyPart = partsTableView.getSelectionModel().getSelectedItem();
         modifyPartIndex = service.getAllParts().indexOf(modifyPart);
-        
+
         displayScene(event, "/fxml/ModifyPart.fxml");
     }
 
@@ -211,15 +208,15 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyProduct(ActionEvent event) throws IOException {
-        modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
+        Product modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
         modifyProductIndex = service.getAllProducts().indexOf(modifyProduct);
-        
+
         displayScene(event, "/fxml/ModifyProduct.fxml");
     }
 
     /**
      * Ask user for confirmation before exiting
-     * @param event 
+     * @param event
      */
     @FXML
     void handleExit(ActionEvent event) {
@@ -229,17 +226,17 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Exit");
         alert.setContentText("Are you sure you want to exit?");
         Optional<ButtonType> result = alert.showAndWait();
-        if(result.get() == ButtonType.OK) {
-            System.out.println("Ok selected. Program exited");
+        if(result.isPresent() && result.get() == ButtonType.OK) {
+            ApplicationLogger.log(Level.INFO, "Ok selected. Program exited");
             System.exit(0);
         } else {
-            System.out.println("Cancel clicked.");
+            ApplicationLogger.log(Level.INFO,"Cancel clicked.");
         }
     }
 
     /**
      * Gets search text and inputs into lookupPart method to highlight desired part
-     * @param event 
+     * @param event
      */
     @FXML
     void handlePartsSearchBtn(ActionEvent event) {
@@ -249,13 +246,11 @@ public class MainScreenController implements Initializable,Controller {
 
     /**
      * Gets search text and inputs into lookupProduct method to highlight desired product
-     * @param event 
+     * @param event
      */
     @FXML
     void handleProductsSearchBtn(ActionEvent event) {
         String x = productsSearchTxt.getText();
         productsTableView.getSelectionModel().select(service.lookupProduct(x));
     }
-
-
 }
